9.31 Move state to HOCs
