+++ src/components/reviews-form/reviews-form.jsx
@@ -1,7 +1,6 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import {connect} from 'react-redux';
-import * as operations from '../../operations/operations.js';
+import withReviewFormState from '../../hocs/with-review-form-state/with-review-form-state.js';
 
 
 const STARS_QUANTITY = 5;
@@ -19,55 +18,18 @@ const StarValue = {
 };
 
 
-class ReviewsForm extends React.PureComponent {
-  constructor(props) {
-    super(props);
-    this._handleRatingChange = this._handleRatingChange.bind(this);
-    this._handleTextChange = this._handleTextChange.bind(this);
-    this._handleFormSubmit = this._handleFormSubmit.bind(this);
-    this.enableForm = this.enableForm.bind(this);
-    this.clearForm = this.clearForm.bind(this);
-    this.state = {
-      rating: 0,
-      text: ``,
-      isFormDisabled: false,
-    };
-  }
+const ReviewsForm = ({formState, onRatingChange, onTextChange, onFormSubmit}) => {
+  const {rating, text, isFormDisabled} = formState;
+  const isSubmitAllowed = rating > 0 && text.length >= TextLength.MIN && text.length <= TextLength.MAX;
 
-  enableForm() {
-    this.setState({isFormDisabled: false});
-  }
-
-  clearForm() {
-    this.setState({rating: 0, text: ``});
-  }
-
-  _handleRatingChange({target}) {
-    this.setState({rating: +target.value});
-  }
-
-  _handleTextChange({target}) {
-    this.setState({text: target.value});
-  }
-
-  _handleFormSubmit(evt) {
-    evt.preventDefault();
-    const {offerId, postComment} = this.props;
-    const {rating, text} = this.state;
-    const comment = {rating, comment: text};
-    this.setState({isFormDisabled: true});
-    postComment(comment, offerId, this.enableForm, this.clearForm);
-  }
-
-  _renderStars() {
-    const {rating, isFormDisabled} = this.state;
+  const renderStars = () => {
     const result = [];
 
     for (let i = STARS_QUANTITY; i > 0; i -= 1) {
       result.push(
           <React.Fragment key={`star` + i}>
             <input
-              onChange={this._handleRatingChange}
+              onChange={onRatingChange}
               checked={rating === i}
               className="form__rating-input visually-hidden"
               name="rating"
@@ -86,47 +48,43 @@ class ReviewsForm extends React.PureComponent {
     }
 
     return result;
-  }
+  };
 
-  render() {
-    const {rating, text, isFormDisabled} = this.state;
-    const isSubmitAllowed = rating > 0 && text.length >= TextLength.MIN && text.length <= TextLength.MAX;
 
-    return (
-      <form onSubmit={this._handleFormSubmit} className="reviews__form form" action="#" method="post">
-        <label className="reviews__label form__label" htmlFor="review">Your review</label>
-        <div className="reviews__rating-form form__rating">
-          {this._renderStars()}
-        </div>
-        <textarea
-          value={text}
-          onChange={this._handleTextChange}
-          className="reviews__textarea form__textarea"
-          id="review"
-          name="review"
-          placeholder="Tell how was your stay, what you like and what can be improved"
-          disabled={isFormDisabled}
-        />
-        <div className="reviews__button-wrapper">
-          <p className="reviews__help">
-            To submit review please make sure to set <span className="reviews__star">rating</span> and describe your stay with <b className="reviews__text-amount">minimum 50 and maximum 300 characters</b>.
-          </p>
-          <button className="reviews__submit form__submit button" type="submit" disabled={!isSubmitAllowed}>Submit</button>
-        </div>
-      </form>
-    );
-  }
-}
-
-ReviewsForm.propTypes = {
-  offerId: PropTypes.number.isRequired,
-  postComment: PropTypes.func.isRequired,
+  return (
+    <form onSubmit={onFormSubmit} className="reviews__form form" action="#" method="post">
+      <label className="reviews__label form__label" htmlFor="review">Your review</label>
+      <div className="reviews__rating-form form__rating">
+        {renderStars()}
+      </div>
+      <textarea
+        value={text}
+        onChange={onTextChange}
+        className="reviews__textarea form__textarea"
+        id="review"
+        name="review"
+        placeholder="Tell how was your stay, what you like and what can be improved"
+        disabled={isFormDisabled}
+      />
+      <div className="reviews__button-wrapper">
+        <p className="reviews__help">
+          To submit review please make sure to set <span className="reviews__star">rating</span> and describe your stay with <b className="reviews__text-amount">minimum 50 and maximum 300 characters</b>.
+        </p>
+        <button className="reviews__submit form__submit button" type="submit" disabled={!isSubmitAllowed}>Submit</button>
+      </div>
+    </form>
+  );
 };
 
-const mapDispatchToProps = (dispatch) => ({
-  postComment(commentData, offerId, enableForm, clearForm) {
-    dispatch(operations.postComment(commentData, offerId, enableForm, clearForm));
-  },
-});
+ReviewsForm.propTypes = {
+  formState: PropTypes.shape({
+    rating: PropTypes.number.isRequired,
+    text: PropTypes.string.isRequired,
+    isFormDisabled: PropTypes.bool.isRequired,
+  }).isRequired,
+  onRatingChange: PropTypes.func.isRequired,
+  onTextChange: PropTypes.func.isRequired,
+  onFormSubmit: PropTypes.func.isRequired,
+};
 
-export default connect(null, mapDispatchToProps)(ReviewsForm);
+export default withReviewFormState(ReviewsForm);
